% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/txGeneNetwork-package.R
\docType{package}
\name{txGeneNetwork-package}
\alias{txGeneNetwork}
\alias{txGeneNetwork-package}
\title{txGeneNetwork: Workflow for constructing Pathway-Gene-Transcript expression networks}
\description{
An R / Bioconductor workflow on how to construct Pathway-Gene-Transcript network from transcript expression data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://villabioinfo.github.io/txGeneNetwork}
  \item \url{https://github.com/villabioinfo/txGeneNetwork}
  \item Report bugs at \url{https://github.com/villabioinfo/txGeneNetwork/issues}
}

}
\author{
\strong{Maintainer}: Lucio Rezende Queiroz \email{luciorqueiroz@ufmg.br} (\href{https://orcid.org/0000-0002-6090-1834}{ORCID})

Authors:
\itemize{
  \item Izabela Mamede Costa Andrade da Conceição \email{iza.mamede@gmail.com} (\href{https://orcid.org/0000-0002-0707-5588}{ORCID})
}

}
\keyword{internal}
